refactor: 생성 Dto를 한 파일에서 관리하도록 수정
