+++ src/main/kotlin/com/example/multidatasourceexample/service/dto/CreateOrderDto.kt
@@ -4,3 +4,13 @@ data class CreateOrderDto(
     val memberId: Long,
     val createOrderItemsDto: CreateOrderItemsDto,
 )
+
+data class CreateOrderItemsDto(
+    val items: List<CreateOrderItemDto>,
+)
+
+data class CreateOrderItemDto(
+    val category: String,
+    val itemName: String,
+    val purchasePrice: Float,
+)
